Refactor nabu to use interfaces that can be tested without global viper config
