+++ internal/graph/ldproc.go
@@ -1,11 +1,12 @@
 package graph
 
 import (
+	"net/http"
+	"os"
+
 	"github.com/piprate/json-gold/ld"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
-	"net/http"
-	"os"
 )
 
 // ContextMapping holds the JSON-LD mappings for cached context
