+++ internal/graph/mock_helpers_test.go
@@ -1,4 +1,4 @@
-package testHelpers
+package graph
 
 import (
 	"context"
