+++ internal/jsonld/processor.go
@@ -1,0 +1,29 @@
+package graph
+
+import (
+	"encoding/json"
+	"fmt"
+
+	log "github.com/sirupsen/logrus"
+	"github.com/spf13/viper"
+)
+
+// JSONLDToNQ takes JSON-LD and convets to nqquads (or ntriples if no graph?)
+func JSONLDToNQ(v1 *viper.Viper, jsonld string) (string, error) {
+	proc, options := JLDProc(v1)
+
+	var myInterface interface{}
+	err := json.Unmarshal([]byte(jsonld), &myInterface)
+	if err != nil {
+		log.Println("Error when transforming JSON-LD document to interface:", err)
+		return "", err
+	}
+
+	triples, err := proc.ToRDF(myInterface, options)
+	if err != nil {
+		log.Println("Error when transforming JSON-LD document to RDF:", err)
+		return "", err
+	}
+
+	return fmt.Sprintf("%v", triples), err
+}
