+++ internal/services/bulk/bulkLoader.go
@@ -8,6 +8,7 @@ import (
 
 	"nabu/internal/objects"
 	"nabu/pkg/config"
+
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
