+++ testHelpers/readme.md
@@ -1,0 +1,1 @@
+the `testHelpers` package contains functions which are used for testing nabu but are not needed for the cli
